# openapi.yaml
openapi: 3.0.3
info:
  title: NeXFlowX Core Banking API
  version: 3.1.0
  description: Motor de Orquestração, Core Bank (In/Pending/Available) e Multi-Tenant SaaS.
servers:
  - url: https://api-core.nexflowx.tech/api/v1
    description: Fintech Core Bank Environment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /auth/login:
    post:
      summary: Autenticação (Suporta Legacy plain-text e Bcrypt)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Retorna Token JWT e role do utilizador
...